# billing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Billing service for popugs.",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/account": {
            "get": {
                "security": [
                    {
                        "OAuth2Implicit": [
                            "read"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Returns user's, account data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Account"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/income": {
            "get": {
                "security": [
                    {
                        "OAuth2Implicit": [
                            "admin",
                            "manager"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Returns company's daily income",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pay": {
            "post": {
                "security": [
                    {
                        "OAuth2Implicit": [
                            "admin",
                            "manager"
                        ]
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Makes payments to users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Account": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.HistoryItem"
                    }
                }
            }
        },
        "model.HistoryItem": {
            "type": "object",
            "properties": {
                "balance_change": {
                    "type": "number"
                },
                "task_description": {
                    "type": "string"
                },
                "task_reward": {
                    "type": "number"
                },
                "task_title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "http://localhost:3000/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "manager": " Grants read and write access to managers information",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    },
    "x-extension-openapi": {
        "example": "value in a json format"
    }
}